refactor: API 요청 처리 클래스 역할 분리
