+++ src/main/java/webserver/APIHandler.java
@@ -18,7 +18,7 @@ public class APIHandler {
     }
 
     private static void fillMethod() {
-        Method[] methods = APIHandler.class.getDeclaredMethods();
+        Method[] methods = RequestProcessor.class.getDeclaredMethods();
         for (Method method : methods) {
             addMapping(method);
         }
@@ -45,28 +45,6 @@ public class APIHandler {
             throw new RuntimeException("처리할 수 없는 요청입니다.");
         }
         Method method = requestMappings.get(httpMethod).find(path);
-        return (HttpResponse) method.invoke(this, httpRequest);
-    }
-
-    @HandleRequest(path = "/", httpMethod = HttpMethod.GET)
-    public HttpResponse ping(HttpRequest httpRequest) {
-        return HttpResponse.ok("hello world");
-    }
-
-    @HandleRequest(path = "/user/create", httpMethod = HttpMethod.POST)
-    public HttpResponse saveUser(HttpRequest httpRequest) {
-        Map<String, String> parameters = httpRequest.getBody();
-        //TODO : NPE 예외처리
-        String name = parameters.get("name");
-        String email = parameters.get("email");
-        String password = parameters.get("password");
-        String userId = parameters.get("userId");
-
-        if (DataBase.findUserById(userId) != null) {
-            throw new RuntimeException("이미 존재하는 회원입니다.");
-        }
-        DataBase.addUser(new User(userId, password, name, email));
-
-        return HttpResponse.found("http://localhost:8080/index.html");
+        return (HttpResponse) method.invoke(RequestProcessor.getInstance(), httpRequest);
     }
 }
